fix(M5): rename misleading react/vue integration modules to adapter names
react.py and vue.py generated JSON Schema data blobs, not React/Vue
components or server-rendered HTML. The names led users to expect
frontend rendering or component generation.

Changes:
- New json_schema_form_adapter.py with ReactJSONSchemaFormAdapter and
  a clear docstring: "outputs data structures, not components"
- New vue_formulate_adapter.py with VueFormulateAdapter and matching
  docstring
- react.py and vue.py reduced to backward-compat shims (re-export old
  names as aliases for the new classes)
- integration/__init__.py now exports ReactJSONSchemaFormAdapter and
  VueFormulateAdapter in __all__
- Tests updated to import from the new canonical module paths and use
  the new class names
