+++ pydantic_schemaforms/integration/vue_formulate_adapter.py
@@ -1,0 +1,65 @@
+"""Vue Formulate data adapter.
+
+This module generates field configuration and validation rule dicts for
+Vue Formulate. It does NOT render HTML or produce Vue components — it
+outputs data structures that a frontend developer wires into a Vue
+Formulate form.
+"""
+
+from __future__ import annotations
+
+from typing import Any, Dict, List, Union
+
+
+class VueFormulateAdapter:
+    """Produce configuration data for Vue Formulate consumers.
+
+    Outputs field config lists and validation rule dicts that a frontend
+    developer passes to Vue Formulate. No HTML is produced; no Vue
+    components are generated.
+    """
+
+    def generate_form_config(self, form_model) -> List[Dict[str, Any]]:
+        config: List[Dict[str, Any]] = []
+
+        for field_name, field_info in form_model.model_fields.items():
+            field_type = field_info.annotation
+            if hasattr(field_type, '__origin__') and field_type.__origin__ == Union:
+                non_none_types = [t for t in field_type.__args__ if t is not type(None)]
+                if non_none_types:
+                    field_type = non_none_types[0]
+
+            field_config = {'name': field_name, 'label': field_name.replace('_', ' ').title()}
+            lowered = field_name.lower()
+
+            if 'email' in lowered:
+                field_config['type'] = 'email'
+            elif 'password' in lowered:
+                field_config['type'] = 'password'
+            elif field_type is bool:
+                field_config['type'] = 'checkbox'
+            elif field_type in (int, float):
+                field_config['type'] = 'number'
+            else:
+                field_config['type'] = 'text'
+
+            config.append(field_config)
+
+        return config
+
+    def generate_validation_rules(self, form_model) -> Dict[str, List[str]]:
+        rules: Dict[str, List[str]] = {}
+
+        for field_name, field_info in form_model.model_fields.items():
+            field_rules: List[str] = []
+            if field_info.is_required():
+                field_rules.append('required')
+            if 'email' in field_name.lower():
+                field_rules.append('email')
+            if field_rules:
+                rules[field_name] = field_rules
+
+        return rules
+
+
+__all__ = ['VueFormulateAdapter']
